Create smooth dev setup
